requirements.txt
# Core requirements for BetterOCR

# OCR Engines
pytesseract>=0.3.10
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.3

# Computer Vision and ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0

# Optional but recommended
python-magic>=0.4.27  # For file type detection
reportlab>=4.0.0  # For PDF generation if needed